+++ orcid-message-listener/src/main/java/org/orcid/listener/common/LastModifiedMessageProcessor.java
@@ -14,7 +14,9 @@
  *
  * =============================================================================
  */
-package org.orcid.listener.listeners.updated;
+package org.orcid.listener.common;
+
+import java.util.function.Consumer;
 
 import javax.annotation.Resource;
 import javax.xml.bind.JAXBException;
@@ -24,7 +26,6 @@ import org.orcid.jaxb.model.record_rc3.Record;
 import org.orcid.listener.clients.Orcid12APIClient;
 import org.orcid.listener.clients.Orcid20APIClient;
 import org.orcid.listener.clients.S3Updater;
-import org.orcid.listener.common.ExceptionHandler;
 import org.orcid.listener.exception.DeprecatedRecordException;
 import org.orcid.listener.exception.LockedRecordException;
 import org.orcid.utils.listener.LastModifiedMessage;
@@ -35,13 +36,16 @@ import org.springframework.stereotype.Component;
 
 import com.amazonaws.AmazonClientException;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.google.common.cache.RemovalListener;
-import com.google.common.cache.RemovalNotification;
 
+/** Core logic for listeners
+ * 
+ * @author tom
+ *
+ */
 @Component
-public class UpdatedOrcidWorker implements RemovalListener<String, LastModifiedMessage> {
+public class LastModifiedMessageProcessor implements Consumer<LastModifiedMessage>{
 
-    Logger LOG = LoggerFactory.getLogger(UpdatedOrcidWorker.class);
+    Logger LOG = LoggerFactory.getLogger(LastModifiedMessageProcessor.class);
 
     @Value("${org.orcid.persistence.messaging.dump_indexing.enabled}")
     private boolean dumpIndexingEnabled;
@@ -57,25 +61,22 @@ public class UpdatedOrcidWorker implements RemovalListener<String, LastModifiedM
     private S3Updater s3Updater;
     @Resource
     private ExceptionHandler exceptionHandler;
-
+    
     /**
      * Populates the Amazon S3 buckets and updates solr index
      */
-    public void onRemoval(RemovalNotification<String, LastModifiedMessage> removal) {
-        if (removal.wasEvicted()) {
-            LastModifiedMessage m = removal.getValue();
-            LOG.info("Removing " + removal.getKey() + " from UpdatedOrcidCacheQueue '" + m.getLastUpdated() + "' Removal cause " + removal.getCause() );            
-            
+    public void accept(LastModifiedMessage m) {
+            String orcid = m.getOrcid();
             try{
-                String orcid = m.getOrcid();
-                                                
                 // Phase #1: update S3 
                 if(dumpIndexingEnabled) {
                     updateS3_1_2_API(orcid);
                     updateS3_2_0_API(orcid);
-                }                
+                } 
+                
             } catch(LockedRecordException lre) {                
                 try {
+                    LOG.error("Record " + orcid + " is locked");
                     exceptionHandler.handleLockedRecordException(m, lre.getOrcidMessage());
                 } catch (JsonProcessingException | AmazonClientException | JAXBException e1) {
                     LOG.error("Unable to handle LockedRecordException for record " + m.getOrcid(), e1);
@@ -84,6 +85,7 @@ public class UpdatedOrcidWorker implements RemovalListener<String, LastModifiedM
                 }                
             } catch(DeprecatedRecordException dre) {
                 try {
+                    LOG.error("Record " + orcid + " is deprecated");
                     exceptionHandler.handleDeprecatedRecordException(m, dre.getOrcidDeprecated());
                 } catch (JsonProcessingException | AmazonClientException | JAXBException e1) {
                     LOG.error("Unable to handle LockedRecordException for record " + m.getOrcid(), e1);
@@ -91,57 +93,14 @@ public class UpdatedOrcidWorker implements RemovalListener<String, LastModifiedM
                     // Should never happen, since it is already deprecated
                 } 
             } catch(Exception e) {
+                //something else went wrong fetching record from ORCID and threw a runtime exception
                 LOG.error("Unable to fetch record " + m.getOrcid() + " so, unable to feed nor S3 nor SOLR");
             }
-        }
+     
     }
     
-    /**
-     * Fetch the public OrcidProfile element from the 1.2 API
-     * @param orcid
-     *          The record id
-     * @return public OrcidProfile or null in case an exception happens
-     * @throws LockedRecordException          
-     * */
-    private OrcidMessage fetchPublicProfile(String orcid) throws Exception {
-        try {
-            return orcid12ApiClient.fetchPublicProfile(orcid);
-        } catch(LockedRecordException lre) {
-            LOG.error("Record " + orcid + " is deprecated");
-            throw lre;
-        } catch (Exception e) {
-            // If we can't fetch the record from 1.2, we should throw any
-            // exception, since we will not be able to index SOLR nor feed S3
-            LOG.error("Unable to fetch OrcidProfile (1.2 API) " + orcid, e);
-            throw e;
-        }
-    }
-    
-    /**
-     * Fetch the public Record element from the 2.0 API
-     * @param orcid
-     *          The record id
-     * @return public Record or null in case an exception happens
-     * @throws LockedRecordException          
-     * */
-    private Record fetchPublicRecord(String orcid) throws LockedRecordException {
-        try {
-            return orcid20ApiClient.fetchPublicProfile(orcid);
-        } catch(LockedRecordException lre) {
-            LOG.error("Record " + orcid + " is deprecated");
-            throw lre;
-        } catch(Exception e) {
-            // If we can't fetch the record from 2.0 we can ignore the error for
-            // now, since the only thing that will fail is feeding S3, which is
-            // still a prototype
-            //TODO: What should we do in this case when this actually moves live?
-            LOG.error("Unable to fetch Record (2.0 API) " + orcid, e);
-        } 
-        return null;
-    }
-    
-    private void updateS3_1_2_API(String orcid) throws Exception {
-        OrcidMessage profile = fetchPublicProfile(orcid);
+    private void updateS3_1_2_API(String orcid) throws LockedRecordException, DeprecatedRecordException {
+        OrcidMessage profile = orcid12ApiClient.fetchPublicProfile(orcid);
         // Update API 1.2
         if(profile != null) {
             try {
@@ -153,8 +112,8 @@ public class UpdatedOrcidWorker implements RemovalListener<String, LastModifiedM
         }                
     }
     
-    private void updateS3_2_0_API(String orcid) throws LockedRecordException {
-        Record record = fetchPublicRecord(orcid);
+    private void updateS3_2_0_API(String orcid) throws LockedRecordException, DeprecatedRecordException {
+        Record record = orcid20ApiClient.fetchPublicProfile(orcid);
         // Update API 2.0
         if(record != null) {
             try {
