--- orcid-message-listener/src/test/java/org/orcid/listener/LastModifiedListenerTest.java
@@ -1,21 +1,0 @@
-/**
- * =============================================================================
- *
- * ORCID (R) Open Source
- * http://orcid.org
- *
- * Copyright (c) 2012-2014 ORCID, Inc.
- * Licensed under an MIT-Style License (MIT)
- * http://orcid.org/open-source-license
- *
- * This copyright and license information (including a link to the full license)
- * shall be included in its entirety in all copies or substantial portion of
- * the software.
- *
- * =============================================================================
- */
-package org.orcid.listener;
-
-public class LastModifiedListenerTest {
-
-}
